PR : modify comment controller, service, repository

+++ src/layers/services/comment.service.js
@@ -18,7 +18,7 @@ module.exports = class CommentService {
     } catch(err){
 
       console.log(err);
-      return { success : false, result : `${err.name} : ${err.message}`};
+      return { success : false, message : err.message };
 
     }
   };
@@ -36,7 +36,7 @@ module.exports = class CommentService {
     } catch(err){
 
       console.log(err);
-      return { success : false, result : `${err.name} : ${err.message}`};
+      return { success : false, message : err.message };
 
     }
 
@@ -46,7 +46,9 @@ module.exports = class CommentService {
 
     try {
 
-      //const isExistComment = await this.postRepository.isExistComment()
+      const isExistComment = await this.commentRepository.isExistComment(commentId);
+      if (isExistComment === null) throw new Error('해당 댓글이 존재하지 않습니다');
+
       const result = await this.commentRepository.updateComment(commentId, text);
       if (result.success === false) throw new Error('댓글 수정에 실패하엿습니다.');
       return result;
@@ -55,7 +57,7 @@ module.exports = class CommentService {
     } catch(err){
 
       console.log(err);
-      return { success : false, result : `${err.name} : ${err.message}`};
+      return { success : false, message : err.message };
 
     }
 
@@ -64,7 +66,9 @@ module.exports = class CommentService {
 
     try {
 
-      //const isExistComment = await this.postRepository.isExistComment()
+      const isExistComment = await this.commentRepository.isExistComment(commentId);
+      if (isExistComment === null) throw new Error('해당 댓글이 존재하지 않습니다.');
+
       const result = await this.commentRepository.deleteComment(commentId);
       if (result.success === false) throw new Error('댓글 삭제에 실패하엿습니다.');
       return result;
@@ -72,7 +76,7 @@ module.exports = class CommentService {
     } catch(err){
 
       console.log(err);
-      return { success : false, result : `${err.name} : ${err.message}`};
+      return { success : false, message : err.message };
 
     }
